Floating point checks are now using appropriate tolerance type

+++ MathLib/include/math/ApproximatelyEqualTo.hpp
@@ -29,12 +29,12 @@ inline bool approximately_equal_to(float input, float near_to, float tolerance =
     return std::abs(near_to - input) <= tolerance;
 }
 
-inline bool approximately_equal_to(double input, double near_to, float tolerance = 0.0002f)
+inline bool approximately_equal_to(double input, double near_to, double tolerance = 0.0002)
 {
     return std::abs(near_to - input) <= tolerance;
 }
 
-inline bool approximately_equal_to(long double input, long double near_to, float tolerance = 0.0002f)
+inline bool approximately_equal_to(long double input, long double near_to, long double tolerance = 0.0002l)
 {
     return std::abs(near_to - input) <= tolerance;
 }
